(GH-194) Add rule and tests for markdown headings
To validate that any headings used within the Description field of the
nuspec file contains valid markdown syntax.  Before the update to the
chocolatey.org website, it would correctly render a heading that used:

However, this is accurately against the markdown spec, and instead
should be written as follows:

This new rule, will provide a Guideline in the package valdiator output
to suggest that the correct syntax be used.