(GH-200) Add unit tests for edge case URL's
It has been found that there are some URL's that are valid but that fail
validation.  These came up as part of this issue, so they have been
added as specific test cases to prevent regressions.  It could be that
over time these URL's change to no longer test what they are meant for,
but for now, they are a good indicator of the problems.