(GH-204) Prevent exception from bubbling up
When a non http/https URL is passed into the url_is_valid function, an
exception is thrown when the WebRequest.Create is called, which wasn't
then caught in the catch, since it was looking for a WebException.
Switch to putting everything in the try, and catch a generic exception,
so that nothing bubles up.