(GH-212) Add CookieContainer to HttpWebRequest
It is possible that an exception can be thrown when requesting a URL
which states "Too many automatic redirects were attempted".  It has
been found that this is caused due to there not being a CookieContainer
for the request in order to capture cookies that are sent with the
response.