(GH-202) Add handling for Protocol error
There is an edge case where a valid URL returns a 503 error code, and
then moves to the site when loaded in the browser.  When loading from
the command line, this isn't handled.  Add a special case for this.
This isn't ideal, but likely not many URL's that are in this position.