(GH-222) Fix 404 response for valid URLs
When setting the UserAgent to something that is specific to Package
Validator, it results in a 404 response.  When using a UserAgent
simliar to what is used when testing URLs from Chrome, makes things
work again. We spoke about this before, and as a result, going to use a
non specific User-Agent string.