(GH-225) Add special handling for Akamai
Simply put, we have no "easy" way of handling the order and case of the
headers that go into the HttpWebRequest.  As such, we will add special
handling for Akamai, such that if the error is a 403, and the Server it
is being returned from is Akamai, allow this to go through validation.
It isn't pretty, but it is really the best that we have at the minute.