(GH-234) Slow down checking collection of URL's
This was added as a result of this issue:
https://github.com/chocolatey/package-validator/issues/234
It isn't pretty, but sending too many requests to GitHub in a row, and
likely other sites, triggers 429 responses, which fails the validation
process.  When looping through a collection of URL's, like those
contained within a set of Release Notes for a package, add a delay
between each check.