(GH-235) Switch to .Net Framework 4.5
This will allow switching to use HttpClient, rather than HttpWebRequest.
Using the WebRequest is proving to have a number of problems, including
the fact that Connection header is always set to Keep-Alive when it
should be keep-alive.  This causes a problem when validating URL's using
Akamai as they "have" to have the correct casing, otherwise URL's are
not validated.

HttpClient is only available starting with .Net 4.5, so have to make the
bump.